diar_sortformer_cache: split framework into espnet3/systems/diar (PR #6459)
Per maintainer review (@sw005320: "new code under espnet3, not espnet2";
@Masao-Someki), move the reusable Sortformer framework out of the recipe into
espnet3/systems/diar and keep only recipe glue in egs3/diar_sortformer_cache:

- espnet3/systems/diar/{system.py (DiarizationSystem), task.py
  (SortformerDiarizationTask)}, metrics/der.py (DER), and a sortformer/ model
  subpackage (model, encoders, cache modules, loss, sliding-window attention,
  longform, inference_bin, converters).
- test -> test/espnet3/systems/diar/test_sortformer.py.
- recipe keeps conf/, dataset/, src/ (data_prep, inference output), run.py,
  path.sh, readme. conf task:/_target_ and run.py now point at
  espnet3.systems.diar.{task,system,metrics.der,sortformer.inference_bin};
  task/system relative imports redirected to the .sortformer subpackage.

Nothing lives under espnet2 anymore. The espnet3 asr/base framework is
untouched. Verified: imports + model build via espnet3.systems.diar.task,
unit tests, and black all pass.
