+++ espnet3/systems/diar/sortformer/transformer_encoder.py
