+++ test/espnet3/systems/diar/test_sortformer.py
@@ -1,10 +1,11 @@
 import torch
-from sortformer.fastconformer_encoder import FastConformerEncoder
-from sortformer.model import ESPnetSortformerModel
-from sortformer.preprocessor import MelSpectrogramPreprocessor
-from sortformer.sort_loss import get_ats_targets, get_pil_targets
-from sortformer.sortformer_modules import SortformerModules
-from sortformer.transformer_encoder import TransformerEncoder
+
+from espnet3.systems.diar.sortformer.fastconformer_encoder import FastConformerEncoder
+from espnet3.systems.diar.sortformer.model import ESPnetSortformerModel
+from espnet3.systems.diar.sortformer.preprocessor import MelSpectrogramPreprocessor
+from espnet3.systems.diar.sortformer.sort_loss import get_ats_targets, get_pil_targets
+from espnet3.systems.diar.sortformer.sortformer_modules import SortformerModules
+from espnet3.systems.diar.sortformer.transformer_encoder import TransformerEncoder
 
 
 def _tiny_model(num_spk=4):
@@ -101,7 +102,7 @@ def test_ats_targets_sorted_by_arrival():
 
 
 def test_eight_speaker_loss_runs():
-    from sortformer.sort_loss import SortformerHybridLoss
+    from espnet3.systems.diar.sortformer.sort_loss import SortformerHybridLoss
 
     torch.manual_seed(0)
     b, t, s = 2, 50, 8
@@ -118,11 +119,11 @@ def test_eight_speaker_loss_runs():
 def test_sliding_window_matches_full_attention():
     """Chunked band kernel == full rel-pos attention when window >= seq length."""
     import torch
-    from sortformer.fastconformer_encoder import (
+    from espnet3.systems.diar.sortformer.fastconformer_encoder import (
         RelPositionalEncoding,
         RelPositionMultiHeadAttention,
     )
-    from sortformer.sliding_window_attention import (
+    from espnet3.systems.diar.sortformer.sliding_window_attention import (
         LocalAttRelPositionalEncoding,
         RelPositionLocalAttention,
     )
@@ -169,10 +170,10 @@ def test_local_attention_encoder_streaming_runs():
 def test_transformer_sliding_window_matches_full():
     """Standard band kernel == full transformer attention when window >= length."""
     import torch
-    from sortformer.sliding_window_attention import (
+    from espnet3.systems.diar.sortformer.sliding_window_attention import (
         TransformerLocalAttention,
     )
-    from sortformer.transformer_encoder import (
+    from espnet3.systems.diar.sortformer.transformer_encoder import (
         TransformerMultiHeadAttention,
     )
